fix: verify signature before replica probe, use ReplicationOp enum
- P1: Move ML-DSA signature verification (step 6) before replica store
  lookup (step 7) to prevent timing oracle — unauthenticated peers
  could probe which (issuer, cid) pairs exist by observing fast-reject
  vs slow-reject paths. Single retrieve() call eliminates double clone.
- P2: Replace raw op u8 with ReplicationOp enum in RuntimeEvent for
  compile-time exhaustiveness checking.
- P2: Document upstream authentication invariant for discovery pubkey
  cache insertion (V1 lacks pubkey→hash binding validation).

+++ crates/harmony-node/src/event_loop.rs
@@ -604,7 +604,7 @@ pub async fn run(
                                         if let Some(contact) = runtime.contact_store().find_by_tunnel_node_id(node_id) {
                                             runtime.push_event(RuntimeEvent::ReplicaReceived {
                                                 peer_identity: contact.identity_hash,
-                                                op: rep_msg.op as u8,
+                                                op: rep_msg.op,
                                                 cid: rep_msg.cid,
                                                 data: rep_msg.payload,
                                             });
